synchronous position close with latest price
